refactor: project cleanup.
