+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = ["apps/desktop/src-tauri", "core", "crates/fs"]
+members = ["adapters/file_system", "apps/desktop/src-tauri", "core"]
