+++ adapters/file_system/Cargo.toml
@@ -1,9 +1,11 @@
 [package]
-name = "md_fs"
+name = "pf_file_system"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+thiserror = "1.0.63"
 rfd = "0.14.1"
+pf_core = { path = "../../core" }
