+++ apps/desktop/src-tauri/Cargo.toml
@@ -25,11 +25,11 @@ opt-level = "z"
 strip = true
 
 [dependencies]
+pf_core = { path = "../../../core" }
+pf_file_system = { path = "../../../adapters/file_system" }
 tauri = { version = "2.0.0-rc", features = [] }
 tauri-plugin-shell = "2.0.0-rc"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-md_core = { path = "../../../core" }
-md_fs = { path = "../../../crates/fs" }
 tokio = { version = "1", features = ["full"] }
 tauri-plugin-dialog = "2.0.0-rc"
