+++ core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "md_core"
+name = "pf_core"
 version = "0.1.0"
 edition = "2021"
