--- core/src/domain.rs
@@ -1,1 +1,0 @@
-pub mod editor;
