+++ core/src/editor.rs
