--- core/src/outbound.rs
@@ -1,2 +1,0 @@
-pub mod data_store;
-pub mod file_system;
